requirements.txt
requests
numpy
torch
Pillow